+++ packages/satoshi-kit/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @midl-xyz/satoshi-kit
 
+## 1.0.2
+
+### Patch Changes
+
+- fix: satoshi kit export
+- Updated dependencies
+  - @midl-xyz/midl-js-react@1.0.1
+  - @midl-xyz/midl-js-core@1.0.1
+
 ## 1.0.1
 
 ### Patch Changes